backend: update - split all controllers and services into individual classes.
