+++ server/railway-ticketing-service/src/main/java/org/project/railwayticketingservice/service/ReservationService.java
@@ -1,14 +1,17 @@
 package org.project.railwayticketingservice.service;
 
 import lombok.RequiredArgsConstructor;
-import org.project.railwayticketingservice.dto.app.request.*;
+import org.project.railwayticketingservice.dto.app.request.GetTrainScheduleRequest;
+import org.project.railwayticketingservice.dto.app.request.NewReservationRequest;
 import org.project.railwayticketingservice.dto.app.response.AppResponse;
-import org.project.railwayticketingservice.dto.app.response.NewTrainResponse;
 import org.project.railwayticketingservice.dto.app.response.ReservationResponse;
 import org.project.railwayticketingservice.dto.app.response.TrainScheduleResponse;
 import org.project.railwayticketingservice.entity.*;
 import org.project.railwayticketingservice.exception.RtsException;
-import org.project.railwayticketingservice.repository.*;
+import org.project.railwayticketingservice.repository.PassengerRepository;
+import org.project.railwayticketingservice.repository.ReservationRepository;
+import org.project.railwayticketingservice.repository.ScheduleRepository;
+import org.project.railwayticketingservice.repository.ScheduleSeatRepository;
 import org.project.railwayticketingservice.util.Utilities;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,19 +19,16 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class AppService {
+public class ReservationService {
 
+    private final ReservationRepository reservationRepository;
     private final PassengerRepository passengerRepository;
     private final ScheduleRepository scheduleRepository;
     private final ScheduleSeatRepository scheduleSeatRepository;
-    private final ReservationRepository reservationRepository;
-    private final TrainRepository trainRepository;
-    private final StationRepository stationRepository;
     private final Utilities utilities;
 
     public ResponseEntity<ReservationResponse> createReservation(NewReservationRequest request) {
@@ -78,13 +78,13 @@ public class AppService {
         scheduleSeatRepository.save(seat);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(
-            ReservationResponse.builder()
-                    .reservationId(reservation.getId())
-                    .train(reservation.getSchedule().getTrain().getName())
-                    .seatNumber(seat.getLabel())
-                    .time(Time.fromLocalDateTime(reservation.getSchedule().getDepartureTime()))
-                    .origin(reservation.getSchedule().getOrigin().toString())
-                    .build()
+                ReservationResponse.builder()
+                        .reservationId(reservation.getId())
+                        .train(reservation.getSchedule().getTrain().getName())
+                        .seatNumber(seat.getLabel())
+                        .time(Time.fromLocalDateTime(reservation.getSchedule().getDepartureTime()))
+                        .origin(reservation.getSchedule().getOrigin().toString())
+                        .build()
         );
 
     }
@@ -143,7 +143,7 @@ public class AppService {
             reservationRepository.save(reservation);*/  // attempt at freeing up reserved seats
             reservationRepository.delete(reservation);
             return ResponseEntity.status(HttpStatus.NO_CONTENT).body(AppResponse.builder()
-                            .message("reservation successfully deleted")
+                    .message("reservation successfully deleted")
                     .build());
         } throw new RtsException(404, "Reservation not found!");
     }
@@ -157,7 +157,7 @@ public class AppService {
         if (!reservations.isEmpty()) {
             reservationRepository.deleteAll(reservations);
             return ResponseEntity.status(HttpStatus.NO_CONTENT).body(AppResponse.builder()
-                            .message("reservations successfully deleted!")
+                    .message("reservations successfully deleted!")
                     .build());
         } throw new RtsException(404, "Reservations not found!");
     }
@@ -203,144 +203,5 @@ public class AppService {
                         )
                         .toList()
         );
-    }
-
-    // admin-specific method
-    public ResponseEntity<AppResponse> createSchedule(ScheduleCreationRequest request) {
-        Train train = trainRepository.findTrainByName(request.train());
-
-        if (train != null) {
-            // check if the schedule actually exists: name -> origin -> departure time
-            if (train.getName().equals(request.train().strip())) {
-
-                for (Schedule schedule : train.getSchedules()) {
-                    // check for station of origin
-                    if (request.origin().equals(schedule.getOrigin().toString())) {
-                        // check for departure times
-                        if (request.departure().getLocalDateTime().equals(schedule.getDepartureTime())) {
-                            throw new RtsException(409, "there is already a schedule fixed for this period.");  // try another train or time?
-                        }
-                    }
-                }
-            }
-
-            // convert to station
-            Station origin = stationRepository.getStationByName(request.origin());
-            Station destination = stationRepository.getStationByName(request.destination());
-
-            if (origin != null && destination != null) {
-
-                Schedule schedule = Schedule.builder()
-                        .train(train)
-                        .currentCapacity(train.getCapacity())
-                        .isFull(false)
-                        .origin(origin)
-                        .destination(destination)
-                        .departureTime(request.departure().getLocalDateTime())
-                        .arrivalTime(request.arrival().getLocalDateTime())
-                        .build();
-
-                scheduleRepository.save(schedule);
-                System.out.println("schedule successfully created");
-
-                // generating seats for schedule
-                utilities.generateSeatsForSchedule(schedule);
-                System.out.println("seats generated");
-
-                return ResponseEntity.status(HttpStatus.CREATED).body(AppResponse.builder()
-                        .message("schedule successfully created.")
-                        .build());
-            } throw new RtsException(400, "Please input a valid station");
-
-        } else {
-            throw new RtsException(400, "Schedule creation failed! No such train!");
-        }
-    }
-
-    // admin-specific method
-    public ResponseEntity<NewTrainResponse> createNewTrain(NewTrainRequest newTrainRequest) {
-        if (!trainRepository.existsByName(newTrainRequest.name())) {
-            Train train = Train.builder()
-                    .name(newTrainRequest.name())
-                    .capacity(Long.valueOf(newTrainRequest.capacity().strip()))
-                    .schedules(null)    // no schedules for now for new train
-                    .build();
-            trainRepository.save(train);
-            System.out.println("train successfully created");
-            return ResponseEntity.status(HttpStatus.CREATED).body(
-                    NewTrainResponse.builder()
-                            .trainId(train.getId().toString())
-                            .trainName(train.getName())
-                            .capacity(train.getCapacity().toString())
-                            .build()
-            );
-
-        } throw new RtsException(409, "train name already exists!");
-    }
-
-    public ResponseEntity<NewTrainResponse> getTrain(String id) {
-        Train train = trainRepository.findTrainById(Long.getLong(id));
-
-        if (train != null) {
-            return ResponseEntity.status(HttpStatus.OK).body(
-                    NewTrainResponse.builder()
-                            .trainId(train.getId().toString())
-                            .trainName(train.getName())
-                            .capacity(train.getCapacity().toString())
-                            .build()
-            );
-        } throw new RtsException(404, "Train not found!");
-    }
-
-    public ResponseEntity<TrainScheduleResponse> getTrainSchedule(String id) {
-        Schedule schedule = scheduleRepository.findScheduleById(id);
-
-        if (schedule != null) {
-            return ResponseEntity.status(HttpStatus.OK).body(
-                    TrainScheduleResponse.builder()
-                            .scheduleId(schedule.getId())
-                            .train(schedule.getTrain().getName())
-                            .currentCapacity(schedule.getCurrentCapacity())
-                            .isFull(schedule.isFull())
-                            .origin(schedule.getOrigin().toString())
-                            .destination(schedule.getDestination().toString())
-                            .departureTime(Time.fromLocalDateTime(schedule.getDepartureTime()))
-                            .arrivalTime(Time.fromLocalDateTime(schedule.getArrivalTime()))
-                            .availableSeats(schedule.getEmptySeats().stream()
-                                    .map(ScheduleSeat::getLabel)
-                                    .collect(Collectors.toList()))
-                            .build()
-            );
-        } throw new RtsException(404, "Schedule not found!");
-    }
-
-    public ResponseEntity<TrainScheduleResponse> editTrainSchedule(String id, ScheduleUpdateRequest request) {
-        Schedule schedule = scheduleRepository.findScheduleById(id);
-        Station origin = stationRepository.getStationByName(request.origin());
-        Station destination = stationRepository.getStationByName(request.destination());
-
-        if (schedule != null) {
-
-            // origin
-            if (!Objects.equals(request.origin(), "null") && !Objects.equals(schedule.getOrigin(), origin)) {
-                schedule.setOrigin(origin);
-                System.out.println("updated origin for train " + id);
-                // destination
-            } if (!Objects.equals(request.destination(), "null") && !Objects.equals(schedule.getDestination(), destination)) {
-                schedule.setDestination(destination);
-                System.out.println("updated destination for train " + id);
-            }   // departure
-            if (request.departureTime() != null && !Objects.equals(request.departureTime(), Time.fromLocalDateTime(schedule.getDepartureTime()))) {
-                schedule.setDepartureTime(request.departureTime().getLocalDateTime());
-                System.out.println("updated departure time for train " + id);
-            }   // arrival
-            if (request.arrivalTime() != null && !Objects.equals(request.arrivalTime(), Time.fromLocalDateTime(schedule.getArrivalTime()))) {
-                schedule.setArrivalTime(request.arrivalTime().getLocalDateTime());
-                System.out.println("updated arrival time for train " + id);
-            }
-
-            scheduleRepository.save(schedule);
-
-        } throw new RtsException(404, "Schedule not found!");
     }
 }
